Moving annotations from the reader reducer (#4083)
- Moving annotations from the reader reducer
