+++ client/app/reader/constants.js
@@ -45,29 +45,6 @@ export const SET_DOC_SCROLL_POSITION = 'SET_DOC_SCROLL_POSITION';
 export const JUMP_TO_PAGE = 'JUMP_TO_PAGE';
 export const RESET_JUMP_TO_PAGE = 'RESET_JUMP_TO_PAGE';
 
-export const OPEN_ANNOTATION_DELETE_MODAL = 'OPEN_ANNOTATION_DELETE_MODAL';
-export const CLOSE_ANNOTATION_DELETE_MODAL = 'CLOSE_ANNOTATION_DELETE_MODAL';
-export const REQUEST_DELETE_ANNOTATION = 'REQUEST_DELETE_ANNOTATION';
-export const REQUEST_DELETE_ANNOTATION_SUCCESS = 'REQUEST_DELETE_ANNOTATION_SUCCESS';
-export const REQUEST_DELETE_ANNOTATION_FAILURE = 'REQUEST_DELETE_ANNOTATION_FAILURE';
-export const REQUEST_CREATE_ANNOTATION = 'REQUEST_CREATE_ANNOTATION';
-export const REQUEST_CREATE_ANNOTATION_SUCCESS = 'REQUEST_CREATE_ANNOTATION_SUCCESS';
-export const REQUEST_CREATE_ANNOTATION_FAILURE = 'REQUEST_CREATE_ANNOTATION_FAILURE';
-export const REQUEST_MOVE_ANNOTATION = 'REQUEST_MOVE_ANNOTATION';
-export const REQUEST_MOVE_ANNOTATION_SUCCESS = 'REQUEST_MOVE_ANNOTATION_SUCCESS';
-export const REQUEST_MOVE_ANNOTATION_FAILURE = 'REQUEST_MOVE_ANNOTATION_FAILURE';
-export const START_EDIT_ANNOTATION = 'START_EDIT_ANNOTATION';
-export const UPDATE_ANNOTATION_CONTENT = 'UPDATE_ANNOTATION_CONTENT';
-export const UPDATE_NEW_ANNOTATION_CONTENT = 'UPDATE_NEW_ANNOTATION_CONTENT';
-export const CANCEL_EDIT_ANNOTATION = 'CANCEL_EDIT_ANNOTATION';
-export const REQUEST_EDIT_ANNOTATION = 'REQUEST_EDIT_ANNOTATION';
-export const REQUEST_EDIT_ANNOTATION_SUCCESS = 'REQUEST_EDIT_ANNOTATION_SUCCESS';
-export const REQUEST_EDIT_ANNOTATION_FAILURE = 'REQUEST_EDIT_ANNOTATION_FAILURE';
-export const START_PLACING_ANNOTATION = 'START_PLACING_ANNOTATION';
-export const PLACE_ANNOTATION = 'PLACE_ANNOTATION';
-export const SHOW_PLACE_ANNOTATION_ICON = 'SHOW_PLACE_ANNOTATION_ICON';
-export const STOP_PLACING_ANNOTATION = 'STOP_PLACING_ANNOTATION';
-export const SELECT_ANNOTATION = 'SELECT_ANNOTATION';
 export const SET_PAGE_COORD_BOUNDS = 'SET_PAGE_COORD_BOUNDS';
 export const SET_UP_PAGE_DIMENSIONS = 'SET_UP_PAGE_DIMENSIONS';
 export const SET_PDF_DOCUMENT = 'SET_PDF_DOCUMENT';
@@ -81,7 +58,8 @@ export const SET_SEARCH_INDEX_TO_HIGHLIGHT = 'SET_SEARCH_INDEX_TO_HIGHLIGHT';
 export const SET_DOCUMENT_LOAD_ERROR = 'SET_DOCUMENT_LOAD_ERROR';
 export const CLEAR_DOCUMENT_LOAD_ERROR = 'CLEAR_DOCUMENT_LOAD_ERROR';
 export const SET_LOADED_APPEAL_ID = 'SET_LOADED_APPEAL_ID';
-export const UPDATE_FILTERED_DOC_IDS = 'UPDATE_FILTERED_DOC_IDS';
+export const UPDATE_FILTERED_RESULTS = 'UPDATE_FILTERED_RESULTS';
+export const ASSIGN_DOCUMENTS = 'ASSIGN_DOCUMENTS';
 
 export const SET_OPENED_ACCORDION_SECTIONS = 'SET_OPENED_ACCORDION_SECTIONS';
 export const SET_VIEWING_DOCUMENTS_OR_COMMENTS = 'SET_VIEWING_DOCUMENTS_OR_COMMENTS';
